backend/pyproject.toml
[project]
name = "prediction-trading-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.96.0",
    "yfinance>=1.3.0",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "matplotlib>=3.10.8",
    "PyYAML>=6.0.2",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
]

[project.scripts]
stock-predictor  = "prediction_trading._cli.stock_predictor:main"
automated-trader = "prediction_trading._cli.automated_trader:main"
scan-watchlist   = "prediction_trading._cli.scan_watchlist:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prediction_trading"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "httpx>=0.27.0",
]
